+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added enum types
 - Improved child node type mapping
 - Fixed root type field names
+- Added input object types
 
 ## [0.2.0] - 2017-10-29